04_Django/11_DRF 2/codes/requirements.txt
asgiref==3.7.2
attrs==24.2.0
Django==4.2.8
djangorestframework==3.14.0
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
pytz==2023.3.post1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.2
uritemplate==4.1.1
